Starts sketching out an interface for IWM drives, eliminating a dangling use of `unsigned` as it goes.

+++ Storage/Disk/Controller/DiskController.cpp
@@ -15,7 +15,7 @@ using namespace Storage::Disk;
 Controller::Controller(Cycles clock_rate) :
 		clock_rate_multiplier_(128000000 / clock_rate.as_int()),
 		clock_rate_(clock_rate.as_int() * clock_rate_multiplier_),
-		empty_drive_(new Drive(static_cast<unsigned int>(clock_rate.as_int()), 1, 1)) {
+		empty_drive_(new Drive(clock_rate.as_int(), 1, 1)) {
 	// seed this class with a PLL, any PLL, so that it's safe to assume non-nullptr later
 	Time one(1);
 	set_expected_bit_length(one);
